- Created OS X target. Resolved all build issues but haven't tested yet. - Added BASS OS X libs - Added two new preprocessor macros: IOS and OSX. So you can #ifdef IOS #endif, etc. - Replaced all NSNumber allocations with literals

+++ Sub/ISMSStreamHandler.m
@@ -213,7 +213,7 @@
 	[encoder encodeBool:self.isTempCache forKey:@"isTempCache"];
 	[encoder encodeBool:self.isDownloading forKey:@"isDownloading"];
 	[encoder encodeInt64:self.contentLength forKey:@"contentLength"];
-	[encoder encodeInt32:self.maxBitrateSetting forKey:@"maxBitrateSetting"];
+	[encoder encodeInt32:(int32_t)self.maxBitrateSetting forKey:@"maxBitrateSetting"];
     [encoder encodeBool:self.isEnableRateLimiting forKey:@"isEnableRateLimiting"];
 }
